geotiff tests: consolidate photometric MinIsWhite + internal-only-JPEG cluster (#2445)
Fold four top-level files into unit/test_photometric.py:

- test_miniswhite_nodata_1809.py
- test_miniswhite_writer_roundtrip_1836.py
- test_signed_miniswhite_rejected_2278.py
- test_to_geotiff_allow_internal_only_jpeg_parity.py

Section 1 is MinIsWhite read/write semantics (mask ordering, writer
round-trip, signed-int rejection). Section 2 is the allow_internal_only_jpeg
API tier. They share a home because both exercise the photometric /
codec-tier API surface, not runtime logic. Boundary comment between the
sections restates that for future readers.

Update docs/source/reference/release_gate_geotiff.rst to point the two
rows that cited the deleted parity file at the new home.

PR B of cluster #2425, long-tail epic #2424.
